# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Convertisseur PDF RFCV → XML ASYCUDA",
    "description": "API REST pour convertir les documents RFCV PDF en XML ASYCUDA",
    "version": "1.1.0"
  },
  "paths": {
    "/api/v1/convert": {
      "post": {
        "tags": [
          "Conversion"
        ],
        "summary": "Conversion synchrone PDF → XML ASYCUDA",
        "description": "Upload un PDF RFCV et retourne le XML ASYCUDA immédiatement. **Taux douanier obligatoire** pour le calcul de l'assurance. Rapport de paiement optionnel.",
        "operationId": "convert_pdf_api_v1_convert_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_convert_pdf_api_v1_convert_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConvertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/convert/async": {
      "post": {
        "tags": [
          "Conversion"
        ],
        "summary": "Conversion asynchrone PDF → XML ASYCUDA",
        "description": "Upload un PDF RFCV et retourne un job_id pour récupérer le résultat plus tard. **Taux douanier obligatoire** pour le calcul de l'assurance. Rapport de paiement optionnel.",
        "operationId": "convert_pdf_async_api_v1_convert_async_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_convert_pdf_async_api_v1_convert_async_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConvertAsyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/convert/{job_id}": {
      "get": {
        "tags": [
          "Conversion"
        ],
        "summary": "Status d'un job",
        "description": "Récupère le status et la progression d'un job de conversion",
        "operationId": "get_job_status_api_v1_convert__job_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/convert/{job_id}/result": {
      "get": {
        "tags": [
          "Conversion"
        ],
        "summary": "Résultat d'un job",
        "description": "Récupère le résultat complet d'un job de conversion avec métriques",
        "operationId": "get_job_result_api_v1_convert__job_id__result_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/convert/{job_id}/download": {
      "get": {
        "tags": [
          "Conversion"
        ],
        "summary": "Télécharger le XML",
        "description": "Télécharge le fichier XML généré pour un job",
        "operationId": "download_xml_api_v1_convert__job_id__download_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_convert_pdf_api_v1_convert_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "Fichier PDF RFCV à convertir"
          },
          "taux_douane": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Taux Douane",
            "description": "Taux de change douanier (ex: 573.1390)"
          },
          "rapport_paiement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rapport Paiement",
            "description": "Numéro de rapport de paiement/quittance Trésor Public (ex: 25P2003J). Optionnel - généré après paiement des taxes."
          }
        },
        "type": "object",
        "required": [
          "file",
          "taux_douane"
        ],
        "title": "Body_convert_pdf_api_v1_convert_post"
      },
      "Body_convert_pdf_async_api_v1_convert_async_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "Fichier PDF RFCV à convertir"
          },
          "taux_douane": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Taux Douane",
            "description": "Taux de change douanier (ex: 573.1390)"
          },
          "rapport_paiement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rapport Paiement",
            "description": "Numéro de rapport de paiement/quittance Trésor Public (ex: 25P2003J). Optionnel - généré après paiement des taxes."
          }
        },
        "type": "object",
        "required": [
          "file",
          "taux_douane"
        ],
        "title": "Body_convert_pdf_async_api_v1_convert_async_post"
      },
      "ConversionMetrics": {
        "properties": {
          "items_count": {
            "type": "integer",
            "title": "Items Count"
          },
          "containers_count": {
            "type": "integer",
            "title": "Containers Count"
          },
          "fill_rate": {
            "type": "number",
            "title": "Fill Rate"
          },
          "warnings_count": {
            "type": "integer",
            "title": "Warnings Count"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "xml_valid": {
            "type": "boolean",
            "title": "Xml Valid"
          },
          "has_exporter": {
            "type": "boolean",
            "title": "Has Exporter"
          },
          "has_consignee": {
            "type": "boolean",
            "title": "Has Consignee"
          },
          "processing_time": {
            "type": "number",
            "title": "Processing Time"
          }
        },
        "type": "object",
        "required": [
          "items_count",
          "containers_count",
          "fill_rate",
          "warnings_count",
          "warnings",
          "xml_valid",
          "has_exporter",
          "has_consignee",
          "processing_time"
        ],
        "title": "ConversionMetrics",
        "description": "Métriques d'une conversion",
        "example": {
          "containers_count": 2,
          "fill_rate": 68.5,
          "has_consignee": true,
          "has_exporter": true,
          "items_count": 12,
          "processing_time": 0.64,
          "warnings": [],
          "warnings_count": 0,
          "xml_valid": true
        }
      },
      "ConvertAsyncResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "message",
          "created_at"
        ],
        "title": "ConvertAsyncResponse",
        "description": "Réponse de conversion asynchrone",
        "example": {
          "created_at": "2025-01-15T10:30:00Z",
          "job_id": "conv_abc123xyz",
          "message": "Conversion en cours",
          "status": "pending"
        }
      },
      "ConvertResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "output_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output File"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConversionMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "processing_time": {
            "type": "number",
            "title": "Processing Time"
          }
        },
        "type": "object",
        "required": [
          "success",
          "job_id",
          "filename",
          "message",
          "processing_time"
        ],
        "title": "ConvertResponse",
        "description": "Réponse de conversion synchrone",
        "example": {
          "filename": "DOSSIER_18236.pdf",
          "job_id": "conv_abc123xyz",
          "message": "Conversion réussie",
          "metrics": {
            "containers_count": 2,
            "fill_rate": 68.5,
            "has_consignee": true,
            "has_exporter": true,
            "items_count": 12,
            "processing_time": 0.64,
            "warnings": [],
            "warnings_count": 0,
            "xml_valid": true
          },
          "output_file": "DOSSIER_18236.xml",
          "processing_time": 0.64,
          "success": true
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "processing",
          "completed",
          "failed"
        ],
        "title": "JobStatus",
        "description": "Status d'un job de conversion"
      },
      "JobStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "progress": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "filename",
          "created_at",
          "message"
        ],
        "title": "JobStatusResponse",
        "description": "Réponse de status de job",
        "example": {
          "completed_at": "2025-01-15T10:30:45Z",
          "created_at": "2025-01-15T10:30:00Z",
          "filename": "DOSSIER_18236.pdf",
          "job_id": "conv_abc123xyz",
          "message": "Conversion terminée",
          "progress": 100,
          "status": "completed"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Clé API pour l'authentification. Pour obtenir une clé API, veuillez contacter l'administrateur du service. Consultez la documentation pour plus d'informations sur l'utilisation de l'API."
      }
    }
  }
}